Xcode Project Generation: include Swift Obj-C Generated Header as public header
Summary:
This adds the ability to reference an `apple_library`'s Generated Obj-C Header from other modules using a framework-style import. For example, if you have a module `Foo` which is a dependency of module `Bar`, you can use the import `<Foo/Foo-Swift.h>` from Obj-C code in `Bar`.

This is achieved by making the following changes:

- `SWIFT_OBJC_INTERFACE_HEADER_NAME` build settings is now explicitly controlled. By default, it will take the value of `Module-Swift.h`.
- By default, Xcode will put the generated header in a private `DerivedSources` folder. This is now explicitly controlled through the `DERIVED_FILE_DIR` so that we know where to find the file.
- Finally, an entry is created in the header map so that `<Module/Module-Swift.h>` gets mapped to `$DERIVED_FILE_DIR/$SWIFT_OBJC_INTERFACE_HEADER_NAME` as we control the values for all the variables in that expression.

The new behaviour is only enabled if `apple_library` has a `swift_version` explicitly defined (for backwards compatibility reasons).

Do note that there's a build dependency if an `apple_library` depends on another `apple_library` which includes Swift sources. That's due to the need to generate the Obj-C header and the Swift module metadata. This dependency is not currently expressed in this diff, it will be addressed in a follow-up diff.

Test Plan:
- Automated tests added as part of diff
- Manual test with a sample project

Reviewed By: dinhviethoa

fbshipit-source-id: a784c02