Add build phase name + deployment postprocessing flag
Summary:
This adds the ability to set the name for a build phase and also control the deployment postprocessing. This functionality will be required in a follow-up diff.

Deployment Postprocessing is represented by Xcode as a checkbox titled "Copy Only When Installing" when used for a file copy phase. This is usually triggered if Product -> Archive is chosen in Xcode.

Test Plan: Unit test + tested with Xcode

Reviewed By: dinhviethoa

fbshipit-source-id: 28e5353