Add support for Skylark's load function
Summary:
`load` function is similar to `include_defs` but requires all included
globals to be explicitly stated. Another advantage is that it's a Skylark built-in
function with full documentation available at: https://docs.bazel.build/versions/master/be/functions.html#load
Having support for this, will allow us to gradually start migrating to Skylark
syntax even when Python interpreter is used instead of Skylark.
It's currently not documented in buck because it's currently experimental.

Test Plan: buck test ...

Reviewed By: styurin

fbshipit-source-id: fb9667b