Prototype Skylark build file parser.
Summary:
This is a simplistic version of Skylark build file parser, which does not support
many features, but will be used a base for further work. Currently it supports parsing only
Buck's native rules. Everything else is not implemented yet and will be implemented in
follow-up changes.

Test Plan: buck test ...

Reviewed By: styurin

fbshipit-source-id: 81b13d3