Remove SharedLibraryInterfaceFactory from dependencies of CxxPlatform
Summary:
This is a part in series of changes to isolate cxx toolchain in `cxx.toolchain` package.

This change decouples `SharedLibraryInterfaceFactory` and its rule-related dependencies from `CxxPlatform`.

The refactoring adds a new class `SharedLibraryInterfaceParams` which contains enough information to re-create `SharedLibraryInterfaceFactory` using `SharedLibraryInterfaceFactoryResolver`. `CxxPlatform` depends on `SharedLibraryInterfaceParams` only and code that needs to instantiate a factory has to use `SharedLibraryInterfaceFactoryResolver`.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: d1c9e26