Remove ABI hack from JavacToJarStepFactory
Summary:
Previously we'd used a field to tell `JavacToJarStepFactory` to generate a source ABI
rather than the full jar. When we started sharing the factory objects between the ABI
and full jar rules (in preparation for pipelining them), that required some ugly
synchronization to prevent races. Now that we're passing parameters in via
`CompilerParameters`, we can specify ABI generation there and eliminate this hackiness.

Test Plan: CI

fbshipit-source-id: a97bc82