Change Android compiler output from classes dir to output jar
Summary:
This diff is a followup to #1412 since the IntelliJ Layout Preview can load classes from the compiler output using the output jar path. This makes it so that you do not necessarily need to change the `spool_to_jar` buckconfig option to `intermediates_to_disk` in order to get Layout Preview working.
Closes https://github.com/facebook/buck/pull/1467

Reviewed By: styurin

Pulled By: styurin

fbshipit-source-id: ba16d69