Added a config to have build rule show only if there take longer than a minimum duration
Summary:
Currently, the superconsole is showing all build rules as they build.
The issue is that the lines of logs in the superconsole is changing too quickly and it's hard to read those lines.

This diff will allow to set a config that will define a minimum duration for build rules before they show in the superconsole.
It will make superconsole more stable.
Therefore, the user will be able to read more easily what's going on in the superconsole.

Reviewed By: ttsugriy

fbshipit-source-id: 1c36c12