Ensure Swift libs folder are passed during linking stage
Summary:
When Xcode compiles static Swift libs, it will include linker commands (`LC_LINKER_OPTION`) that will be carried over for the final binary to link to the appropriate Swift overlays and libs. This means that the final binary must be able to locate the Swift libs in the library search path. If an Xcode target includes Swift, Xcode will automatically append the Swift lib folder when invoking the linker. For example, it will insert the following path:

    /Applications/Xcode.app/Contents/Developer/Toolchains/XcodeDefault.xctoolchain/usr/lib/swift/macosx

which can be dynamically specified as:

    $DT_TOOLCHAIN_DIR/usr/lib/swift/$PLATFORM_NAME

Unfortunately, this will not happen if we have a plain `apple_binary` that has Swift deps. So we're manually doing exactly what Xcode does to make sure binaries link successfully if they use Swift directly or transitively.

To extract `LC_LINKER_OPTION` from a static library:

    otool -l libFoo.a

You will be able to see output like:

    Load command 17
         cmd LC_LINKER_OPTION
     cmdsize 32
       count 1
      string #1 -lswiftCoreGraphics

Test Plan: Automated tests part of diff + manual Buck setup test

Reviewed By: VTopoliuk

fbshipit-source-id: f76171b