Weakly cache the manifest RuleKeyAndInputs
Summary:
Due to the async nature of CachingBuildRuleBuilder, there may be many
BuildRules that are in-between two stages that both need the manifest's
RuleKeyAndInputs. If we just stored the RuleKeyAndInputs directly, we
could use too much memory.

This diff introduces a WeakMemoizingSupplier<RuleKeyAndInputs> that is
used in performManifestBasedCacheFetch. This should resolve an issue
where BUCK was using a lot of memory due to many BuildRules computing
RuleKeyAndInputs and then waiting for the manifest fetches (previously
the fetches were synchronous and so we were limited to just however many
threads we were using).

Currently, we still recompute the RuleKeyAndInputs in
finalizingBuildRule. I've added a TODO there, I just didn't want to
muddy fixing the OOM issue with optimizing that call. I'll followup
immediately with a diff that resolves the TODO.

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: 23291f7