Change file-expansion-prefix for test selectors from @ to :
Summary:
`@` symbol file expansion happens in buck prior to subcommand arg parsing.
This means that subcommands which wish to handle file expansion themselves must use
a different prefix char. `:` was chosen because it resembles a buck target reference,
and was not a reserved character in the regex dialect.

Test Plan: new unit tests

Reviewed By: ttsugriy

fbshipit-source-id: b571020