Reapply "Support empty header_path_prefix" patch with fixes.
Summary:
This reverts commit 47434ecb22fd00760234f2839a0ff89d0ca6c376, which was
necessary to fix internal build breakages. This change uses a more generic
algorithm to avoid header duplicates when `header_path_prefix` is empty -
duplicates are allowed as long as they map to the same path. This is
a little sloppy, since we can potentially be doing a bit more work, but
the implementation is much easier to maintain.

Test Plan:
buck test ...
also build internal builds using this version of buck and make sure they can be built now

Reviewed By: dinhviethoa

fbshipit-source-id: 86f9d9c