Plumb TreePath into PostEnterCanonicalizer
Summary:
In order to properly infer types in some cases, we need to be able to get the type of
subexpressions. The only methods available on `Trees` for doing so require a `TreePath` rather
than just a `Tree`, so this commit gets `TreePath`s into all the right places.

Test Plan: CI, integration tests later in the stack

fbshipit-source-id: f6794bc