Failing tests for descriptors and signatures without deps
Summary:
This commit parameterizes `DescriptorFactoryTest` and `SignatureFactoryTest` so they can
run with and without dependencies. Right now they fail without dependencies because that
support has not yet been implemented, so the tests are ignored when running that way.

Test Plan: Remove the `assume` and verify the tests fail because of not being able to infer types.

fbshipit-source-id: e2ade8f