Enable ABI generation to emit error messages
Summary:
We'd like to be able to supply useful error messages when code is not up to standard for source ABI
generation. This commit implements the annotation processor `Messager` interface (which is available
even outside the compiler's class loader) in terms of `Trees` (which is not) so that our compiler-agnostic
ABI generation code can emit such messages.

Test Plan: Use it

fbshipit-source-id: 58067d7