Implement DefaultClassInfo
Summary:
This adds an implementation of DefaultClassInfo that supports a small
set of things. There's some TODOs for implementing other stuff (Pair,
Either, ImmutableSortedMap, ...).

With this, we can begin converting some (simpler) buildrules to
ModernBuildRule.

The implementation is pretty similar to AlterRuleKey stuff, but it just
always operates on all fields of a Buildable and all of its
superclasses.

There's a bunch of preconditions to reject things that we don't want to
allow in Buildables.

Test Plan: Added tests for correct behavior and for correct failures

Reviewed By: yiding

fbshipit-source-id: d1c7552