Implement AbstractAsynchronousCache multifetch
Summary:
This makes the abstract cache handle dispatching of fetch/multifetch
based on how many outstanding requests there are.

There's multiFetchLimit() and getConcurrency() methods that implementors
can override to control how many keys should get sent to multifetch. If
multiFetchLimit() < 2 (or if based on concurrency and outstanding
requests we're only going to send a single key), then fetch will be used
instead of multifetch. As multiFetchLimit() is checked on every call, it
provides a method for the cache to disable multiFetch during a build (if
it wants).

Test Plan: CI, added simple test

Reviewed By: ttsugriy

fbshipit-source-id: 2530f9a