Make handling of SettableFuture less brittle
Summary:
AbstractAsynchronousCache should be careful to *always* set a future if
it takes a FetchRequest out of the pending queue (or return the request
to the queue). Currently, if the underlying implementations throw
exceptions in particular ways, then these futures may not be set.

This introduces a ClaimedFetchRequest which helps handle these futures
correctly. It's a closeable that will set the future if it hasn't been
set or rescheduled. It also enforces that setting the
future/rescheduling the request is the last thing that is done to it.

Test Plan: CI, introduce a bad exception (in events.failed() for example) and see that tests fail correctly

Reviewed By: ttsugriy

fbshipit-source-id: 51f32e0