Fix cross-cell project filesystem switch when somerging
Summary:
When fixing up dependencies from somerging, we use the current project filesystem regardless of the underlying target. For targets all in the same cell this is fine, but for purely cross-cell dependencies, this switches the current working directory. We then later get a compilation error because of the switched directory.

And add a unit test for when there is a cross-cell dependency on a library that is merged.

Test Plan:
Add new unit test. And it passes. Fails without the additional change with

    i686-linux-android-g++: error: buck-out/gen/library/I#android-x86,compile-pic-I.c.ofd3ceae6/I.c.o: No such file or directory
    BUILD FAILED: secondary//library:I#android-x86,merge_structure_e5a05f58 failed with exit code 1:
    c++ link
    stderr: i686-linux-android-g++: error: buck-out/gen/library/I#android-x86,compile-pic-I.c.ofd3ceae6/I.c.o: No such file or directory

Reviewed By: dreiss

fbshipit-source-id: c1109b3