Flush streams when forwarding stdout and stderr of subprocess
Summary:
When using an external runner (and some other cases), the output and error
streams of the subprocess are forwarded to buck's stdout and stderr, but the
streams are not flushed. In the event that buck's stdout and stderr is
block-buffered (which can happen if buck is run without buckd and the output is
not a tty), output may not be displayed immediately after being produced by the
underlying process.

I fix this by flushing after each write call.

Test Plan:
Write a script that emulates an external test runner:

  ❯ cat test.sh
  #!/bin/bash

  echo "1. stdout"
  sleep 5
  >&2 echo "2. stderr"
  sleep 5
  echo "3. stdout"

Run by redirecting output stream to files, and tailing those files. Incorrect
execution would emit the lines out of order or at an interval that's not every
5 seconds (in particular, stdout messages are printed at the end
simultaneously). Correct execution would emit lines in-order every 5 seconds.

  ❯ NO_BUCKD=1 ~/buck2/bin/buck test \
      --config test.external_runner=$PWD/test.sh \
      test/com/facebook/buck/util/sha1:sha1 \
      > test.stdout 2> test.stderr & \
      sleep 1 && tail -f test.stdout & sleep 1 && tail -f test.stderr
  ...
  1. stdout
  2. stderr
  3. stdout

Reviewed By: ttsugriy

fbshipit-source-id: e6636f6