Fix FakeClock: Introduce FakeClock.DO_NOT_CARE
Summary:
`FakeClock` is perpetuating a dangerous assumption...that `Clock.nanoTime` is just the nanoseconds
version of `Clock.currentTimeMillis`. The `System` class, after which `Clock` is patterned and with
which `DefaultClock` is implemented, definitely does not behave this way, so our tests are potentially
hiding timestamp-related bugs.

This commit moves all tests that used `new FakeClock(0)` to use a new `FakeClock.DONT_CARE` constant,
in prepration for changing the construction of `FakeClock` to be aware of the difference between
`currentTimeMillis` and `nanoTime`.

This was a straightforward find-replace in IntelliJ.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: b217aad