Fix FakeClock: Switch a few more tests to `DO_NOT_CARE`
Summary:
`FakeClock` is perpetuating a dangerous assumption...that `Clock.nanoTime` is just the nanoseconds
version of `Clock.currentTimeMillis`. The `System` class, after which `Clock` is patterned and with
which `DefaultClock` is implemented, definitely does not behave this way, so our tests are potentially
hiding timestamp-related bugs.

The tests changed in this commit were not as obviously uncaring as those which were initializing
`FakeClock` with `0`, but on closer inspection they do not need any particular timestamp value.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 3f26c10