Fix FakeClock: Move to Immutable, validate instances
Summary:
`FakeClock` is perpetuating a dangerous assumption...that `Clock.nanoTime` is just the nanoseconds
version of `Clock.currentTimeMillis`. The `System` class, after which `Clock` is patterned and with
which `DefaultClock` is implemented, definitely does not behave this way, so our tests are potentially
hiding timestamp-related bugs.

This commit moves `FakeClock` to use immutables, adding a method to check for validity of the timestamp
values, and updates the remaining tests that care about timestamps to use this properly.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 51a68e0