Fix SettableFakeClock
Summary:
`SettableFakeClock` had a similar problem to `FakeClock` in that it didn't prevent people
from setting the clock to an unrealistic state in which `currentTimeMillis` and `nanoTime` were
directly related. This commit reworks `SettableFakeClock` in terms of the new `FakeClock`. As it turns
out, all tests that were being bad in their value setting didn't care about the specific values, so
we just migrated them to `SettableFakeClock.DONT_CARE`.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: dfdf0ef