Support async operations in FileContentsProvider
Summary:
This diff adds support for async file materialization in all implementations of
FileContentsProvider. In the current model, we tried to make use of the fact that there are
multiple threads with file requests, but the problem is that there are not many build threads, but
each build thread usually has (in comparison) a huge number of files it wants to materialize.

This diff adds async operations to the FileContentsProvider interface, which is essential in
removing the bottleneck described above.

Test Plan: CI

Reviewed By: alisdair04

fbshipit-source-id: 05e419e