Support async file materialization in MaterializerProjectFileHashCache.
Summary:
This diff adds support for async file materialization in MaterializerProjectFileHashCache
(now MaterializerDummyFileHashCache). In the current model, we tried to make use of the fact that
there are multiple threads with file requests, but the problem is that there are not many build
threads, but each build thread usually has (in comparison) a huge number of files it wants to
materialize.

This diff adds makes it so that the MaterializerFileHashCache uses the async method to fetch files
from FileContentsProvider. It then stores the Futures in a member map, to keep track of the
materialization. So DistBuildCachingBuildEngineDelegate now issues the requests for file
materialization in onRuleAboutToBeBuilt (via RuleKey computation), but then waits for the
Materializer to finish fetching everything it needed to.

One side-effect of this diff is that MaterializerProjectFileHashCache is no longer able to return
the correct HashCode in the get methods. But this should not be an issue because it is used for the
sole purpose of materializing files.

Test Plan: CI

Reviewed By: alisdair04

fbshipit-source-id: 6911f11