Modify the defaults for source file multi-fetch.
Summary:
For now we're making preload-materialization the default strategy so as to clearly put aside
the time for materializing files. Once we launch stampede into production, we can switch to on-demand
materialization if it performs better.

Test Plan:
Ran the build slave locally and experimented with a lot of different values for these
parameters. Materializing during preload with a max buffer size of 40 performs decently in a large
build (materializing more than 28k files in about 12-13s -- which we can hide by doing in parallel
with action graph creation). On-demand materialization depends on a lot of variables, like the
average number of files requested by the rules, or the highest number of files requested by a single
rule.

Reviewed By: alisdair04

fbshipit-source-id: 6f75328