Use safe values for flags in InnerClasses table
Summary:
The `InnerClasses` table of a `.class` file contains information about any member class
that is referenced from that file. When building source ABIs without deps, some of those
member classes may be inferred types. We cannot know the flags for inferred types, so
we must guess. As it turns out, setting just `ACC_STATIC` is safe in terms of the
actions that the compiler will take based on the flags. See the comment in
`InnerClassesTable` for more info.

Test Plan: New unit tests

Reviewed By: dreiss

fbshipit-source-id: 7011e13