Reduce amount of flickering in superconsole
Summary:
Before:
- It was clearing everything
(when the screen updates at this time, this part of the screen can turn empty)
- Then writing all the new output

After diff:
On each line:
- clear the current line
- write a new line of output
- then, if there are remaining lines from the old output, we clear them

By writing the output in that order, we reduce the chance of accidental empty output.

Test Plan:
```
$ buck build [app]
```

Visually check that there's less flickering using buck with this change.

Reviewed By: ttsugriy

fbshipit-source-id: a0e9003