Have InterfaceScanner look for annotation definitions too
Summary:
Any rule that defines an annotation should be marked with `required_for_source_abi = True`. To facilitate
that, we'll be outputting warnings for any that aren't, much as we do for names that need to be qualified.

This commit enables `InterfaceScanner` to point out annotation definitions. Subsequent commits will
change `InterfaceValidator` to do something about it.

Test Plan: New unit test

Reviewed By: dreiss

fbshipit-source-id: 52938ee