Fix test issue in FakeBinaryBuildRuleBuilder
Summary:
This thing is creating an `Arg` class with a `getDeps` method that doesn't implement `HasDeclaredDeps`
(the interface from which that method usually comes). As a result, during testing the dependencies
declared on one of these guys was not getting picked up through the normal mechanism as a dependency,
but rather as an "extra" dependency by way of an overzealous `TargetNodeFactory`. This commit defines
the `Arg` class correctly, so that the deps come via the right path, on the way to making `TargetNodeFactory`
less zealous.

Test Plan: CI

fbshipit-source-id: 6f93bc7