Plumb required_for_source_abi down to InterfaceValidator
Summary:
Any rule that defines an annotation should be marked with `required_for_source_abi = True`. To facilitate
that, we'll be outputting warnings for any that aren't, much as we do for names that need to be qualified.

`InterfaceValidator` is the thing that outputs such warnings; this commit plumbs `required_for_source_abi`
into it so that it has the necessary information.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 15ce67f