Emit warnings or errors when annotations are defined in targets without required_for_source_abi = True
Summary:
Any rule that defines an annotation should be marked with `required_for_source_abi = True`. To facilitate
that, we'll be outputting warnings for any that aren't, much as we do for names that need to be qualified.

This commit makes the final changes in `InterfaceValidator` to make this happen.

Test Plan: New unit test

Reviewed By: dreiss

fbshipit-source-id: 190dbdb