Implement Skylark filesystem based on ProjectFilesystem
Summary:
This implementation tries to delegate as much as possible to
underlying `ProjectFilesystem` and in cases when it's not possible,
uses Java APIs. Due to extra conversions we might have to consider
caching or using some other mechanism to reduce the number of
allocations, but for now it's fine.

Test Plan: buck test ...

Reviewed By: dinhviethoa

fbshipit-source-id: 82df390