buck project: list dependencies in the same order as how buck builds classpath
Summary: When buck builds classpath, it adds all X's direct dependencies sorted by name, before adding dependency X.  The order of entries in classpath is important when there are duplicated classes, either by design or accident.  This change lists dependencies in module.iml with the same order so that Intellij uses the same order of entries in classpath as buck.

Test Plan:
unit tests
manual test with a project

Reviewed By: styurin

fbshipit-source-id: f7bc430