unset CAML_LD_LIBRARY_PATH in debug script
Summary:
unset CAML_LD_LIBRARY_PATH in debug script

If you have opam installed, then it will set this environment variable
to something like

    ~/.opam/system/lib/stublibs:/usr/lib64/ocaml/stublibs

which confuses the debug script and directs it to use libraries that
buck doesn't know about.

deleting this environment variable or setting it to the empty string is
enough to get the debug script to pick up the correct libraries.

Test Plan:
ran manually modified script with this change and It Worked
On My Machine.

Reviewed By: mroch

fbshipit-source-id: a61a25a