Move android:packageable to android.packageable
Summary:
Moving classes related to `AndroidPackageable` to a standalone package.

This interface is used by other classes outside of `android` package to specify artifacts that need to be packaged in apks. Since it's used by outside classes it's better to keep it away from the rest of android classes.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: ed88377