Stop passing CommandEvent.Started to the Daemon and posting it to its fileEventBus.
Summary:
In D5596766, I made a change that has `ChromeTraceBuildListener` register itself
with the `Daemon`'s `fileEventBus`. The purpose of this change was to make it so
that the `ChromeTraceBuildListener` could get notified about
`WatchmanOverflowEvent`s. This appears to have had an unintended side-effect
where the `ChromeTraceBuildListener` would get notified about the
`CommandEvent.Started` event twice.

Like most `BuckEventListener`s that care about the `CommandEvent.Started` event,
the `ChromeTraceBuildListener` registers itself with the `buckEventBus` in
`Main.java`. However, for some reason that probably made sense at some point in
history but no longer seems to make sense today, the `fileEventBus` was also
a publisher of the `CommandEvent.Started` event in the `Daemon`. This is
particulary odd because it does not publish the `CommandEvent.Finished` event.
Further, the role of the `fileEventBus` appears to be to publish
`WatchmanEvent`s from the `WatchmanWatcher`, so subscribers of `fileEventBus`
should not expect to get non-Watchman events.

Looking through `Daemon`, the `fileEventBus` is registered with the
following subscribers:
* `WatchedFileHashCache`
* `Parser`
* `ActionGraphCache`
* `RuleKeyCacheRecycler`

Looking through those four classes and grepping for `Subscribe`, the only
things that any of those classes care about are implementations of
`WatchmanEvent` (notably `WatchmanPathEvent` and `WatchmanOverflowEvent`). I
cannot find any functionality they would lose by no longer having the
`CommandEvent.Started` event broadcast to them.

Test Plan:
Will see if CI turns up any behavior that relies on the existing behavior.

Previous to this change, I ran:

```
jq '[.[] | select(.args.command_args) ]' buck-out/log/build.trace
```

and saw things like:

```
[
  {
    "cat": "buck",
    "name": "query",
    "ph": "B",
    "pid": 0,
    "tid": 365,
    "ts": 654234590538,
    "tts": 59640,
    "args": {
      "command_args": "owner('%s') pkg/nuclide-clang-rpc/python/references.py"
    }
  },
  {
    "cat": "buck",
    "name": "query",
    "ph": "B",
    "pid": 0,
    "tid": 365,
    "ts": 654234590538,
    "tts": 59640,
    "args": {
      "command_args": "owner('%s') pkg/nuclide-clang-rpc/python/references.py"
    }
  },
  {
    "cat": "buck",
    "name": "query",
    "ph": "E",
    "pid": 0,
    "tid": 365,
    "ts": 654234818684,
    "tts": 65008,
    "args": {
      "command_args": "owner('%s') pkg/nuclide-clang-rpc/python/references.py",
      "daemon": "true"
    }
  }
]
```

Note how the `CommandEvent.Started` event is posted twice (with 100% identical
data), but the `CommandEvent.Finished` event is posted once.

After this change, the begin and end events appeared exactly once.

Reviewed By: ttsugriy

fbshipit-source-id: 9c8e2b5