Start dogfooding skylark parser for buck repo
Summary:
This is the best way to ensure that Skylark parsing continues
to work. Currently there is only one file (third-party/java/bazel/BUCK) that
is Skylark enabled, so the risk is minimal, but with time we should be
able to extend the number of Skylark-enabled BUCK files.

Test Plan: CI

Reviewed By: sbalabanov

fbshipit-source-id: 01dcc47