Fix embedding of Swift runtime on macOS
Summary: This fixes a bug in how the Swift runtime is embedded. The code is only correct for iOS but not for macOS. `AppleBundle` already has a reference to `AbstractAppleBundleDestinations` which specifies the appropriate platform paths, so we just use that instead.

Test Plan: Automated testing. In addition, the new code will result in the path, on iOS, to be equal to `Paths.get("").resolve("Frameworks");` (defined in `AbstractAppleBundleDestinations.java`). I've verified that `Paths.get("").resolve("Frameworks").equals(Paths.get("Frameworks"))`, so behavior will be unchanged on iOS. On macOS, the new path is correctly used (tested in internal prototype).

Reviewed By: dinhviethoa

fbshipit-source-id: 9f1ccf3