WorkThreadTrackingFuture - Futures that track the thread currently computing its value.
Summary:
Introduce WorkThreadTrackingFuture, which tracks the thread currently computing its value.

This future keeps track of the thread currently evaluating it, throwing an error if the thread
tries to call get  on the future while evaluating the future. It also lets a thread query whether
it is currently evaluating the future, and react accordingly

An example of where this is useful.

This is useful when converting a cache to hold futures instead of bare values, in order to
prepare for the cache to be concurrently accessed without duplicating work, where the
computation being cached may query the cache for itself. (Yes this happens all the time
in Buck).

Test Plan: Added test.

Reviewed By: ttsugriy

fbshipit-source-id: 78493b6