MultiThreadedBuildRuleResolver - BuildRuleResolver that allows action graph creation to be parallelized
Summary:
Introduces a BuildRuleResolver implementation that allows action graph creation
to be parallelized.

Using ForkJoinPool/ForkJoinTask, this implementation allows target node
transformations to block on dependencies, so that no duplicates are created,
without rewriting all the code to use CompletionStage/ListenableFuture, and
without spawning unbounded number of worker threads.

One bit of special handling comes from recursive calls by a task to compute its
own BuildTarget. This usually happens via a call chain in the style of
`requireRule(fooTarget) -> createBuildRule(fooTarget) ->
computeIfAbsent(fooTarget)`. To resolve this, if the current thread is working
on a build target, then recursive calls will pretend like it doesn't exist,
instead of waiting, which emulates the old, single threaded behavior and allows
this kind of calls to keep working.

Note that in order to actually achieve parallelism, calls to requireRule must
be issued in parallel. This will be handled in future diffs.

Test Plan: I should probably test some of the new semantics of this build rule resolver here.

Reviewed By: sbalabanov

fbshipit-source-id: 97488bb