Parallelize action graph creation
Summary:
Introduce a config option which allows action graph creation to be
parallelized. The current implementation utilizes the common ForkJoinPool for
simplicity, but I may create a per-command fork join pool to handle this in a
future diff.

Test Plan:
Ran `buck target --show-rulekey` on several targets, and see that they produce
the same results.

Reviewed By: sbalabanov

fbshipit-source-id: a30a758