SwiftPlatform defines list of rpaths
Summary: The rpaths for the dynamic linker should depend on `SwiftPlatform` because they're different on iOS and macOS. This diff makes `SwiftRuntimeNativeLinkable` use `SwiftPlatform` to access the rpaths, rather than hardcoding them. Handling the rpath differences between macOS and iOS would be addressed as a follow-up.

Test Plan: This diff should not change behavior. The pre-existing test `SwiftNativeLinkableTest.java` checks the linker flags which are now being accessed through `SwiftPlatform`.

Reviewed By: dinhviethoa

fbshipit-source-id: 5995628