TreeBackedAnnotatedConstruct.getAnnotation/getAnnotationsByType
Summary:
While the implementation of these in `TreeBackedAnnotatedConstruct` couldn't be more trivial,
there are some gotchas here. Most of them are taken care of by the fact that the types of
annotations and any constant values passed to them (except class literals) must be present
in the classpath even when compiling "without dependencies", else the compiler will error
out before even giving us a chance.

The one case that remains is class literals -- if an annotation processor uses
`getAnnotation/getAnnotationsByType`, any classes passed as class literals to such annotations
must be present on the classpath. This isn't much of hardship; even with deps such classes must
be on *both* the source-code classpath and the annotation processor's classpath, which limits
their number.

Test Plan: New unit tests

Reviewed By: dreiss

fbshipit-source-id: 1c1e3d3