Handle resources_root in project generation
Summary:
IntelliJ project generation right now doesn't generate resource folders for Java code, instead relying directly on the generation of source folders. This results in resources being put in the wrong places if resources_root is different from the Java src_roots option. This change introduces a new type of IjFolder (JavaResourceFolder) which tracks resources_root if it is set and sets the relativeOutputPath of the IntelliJ folders to try and handle this correctly.

Note that it probably can't be fully handled correctly if there are source files and resource files mixed in the same directories, because IntelliJ's granularity only goes to the directory level, while buck can set them at the file level.

I'm not entirely sure if this is the best approach to handling this, would appreciate any suggestions
Closes https://github.com/facebook/buck/pull/1410

Reviewed By: styurin

Pulled By: styurin

fbshipit-source-id: 666ac50