Don't add SDKROOT to framework search paths
Summary:
Not all of Apple's system frameworks have correct modular headers.
Adding the `$SDKROOT` in a `-F` flag triggers an attempt of clang to
create a module for these, which gives unexpected errors.

Omitting the `$SDKROOT` is not an issue when the `-sdk` flag is passed to
`clang`/`swiftc`. When generating a new project in Xcode, it does not
include the `$SDKROOT` variable in `FRAMEWORK_SEARCH_PATHS` by
default.

Related radars:
- http://openradar.appspot.com/33774557
- http://openradar.appspot.com/33795982
Closes https://github.com/facebook/buck/pull/1502

Reviewed By: yiding

Pulled By: yiding

fbshipit-source-id: cbac05a